typechecker: Properly keep track of generic inferences across scopes
Previously we were using an empty or incorrect inference set in many
places, and using TypeId::equals() to check for type compatibility,
which lead to a slew of issues in generic functions and their
instantiations.
This commit should fix most of those, and also allow function
expressions with generic types to actually work.