all: Move `NumericConstant` parsing to the parser
This changes the lexer to output a `Number` token with:
`prefix: LiteralPrefix, number: String, suffix: LiteralSuffix`
instead of
`number: NumericConstant`

The parser then parses this token into a `NumericConstant`

The primary motivation for this is so that the Number token carries the
information required to reconstruct the source for formatting. It also
makes for a better separation of concerns though as this was parsing
that the lexer was doing.

This also changes the lexing slightly in that for example for an octal:
`0o78` would end up as `Garbage` due to the 8 being an invalid digit.
It would now lex as `0o7` and then a number `8`

`0o8` would lex as `0o` and `8`, with `0o` giving an error in the parser
as having no digits.