jakttest: Allow specifying the output in a separate file
Previously the output specification was only a quoted string, for
example:
/// - output: "PASS\n"

If a non quoted string is passed it is now assumed that a filename for
the expected output is being read, for example:
/// - output: output.txt

Where output.txt would contain:
/// Expect: Result

Pass

"Expect: Result" indicates that the file is a result file and shouldn't
be tested/ marked as skipped. Any newlines after this line are ignored
for the comparison.