typechecker: Avoid generic inference of T = T
This happens for calls when we have 2 generic types with the same name
in different scopes.

As in the example where it is a struct `Enumerator<T>`
It has a member `iterator: ArrayIterator<T>`
We make a call `let item: T? = .iterator.next()`
The `T?` returned by `.iterator.next()` is a different type to the `T?`
that is given to item.

Possibly they should be the same type_id or have another method of
resolution that isn't just comparing the names.